decouple dashborad and reservation from the homepage
